add comunication user-booking with rabbit and update notification service

+++ user-service/src/main/java/it/roadies/user_service/services/GamificationService.java
@@ -1,5 +1,5 @@
 package it.roadies.user_service.services;
 
 public interface GamificationService {
-    void addPointsBySpending(String userId, double amountSpent);
+    void addPointsBySpending(String userId, Long amountSpent);
 }
